fix(project-generator): Filter step skips an index
When generating `syndesis.yaml` it's important that the index of the
step is in line with the generated `application.properties` in
`/controllers/src/main/java/io/syndesis/controllers/integration/IntegrationSupport.java`

In the case when the filter step is followed by two or more steps the
nested steps are created with the same context, and thus share the same
index.

This tracks the context that needs to be used to maintain the index.

Fixes #888