chore: release using GitHub Actions
This adds a GitHub Actions workflow to release Syndesis. To trigger a
release one needs to push a tag with name `major.minor.patch`, e.g.
`1.17.0`. Use of a local repository cache in a temporary directory was
removed, as we need to rely on the caching provided by GitHub Actions.
We also use strict checksums in release so that makes it even safer.