fix: no need for junit4-stub in release
When doing a release build with `-DstagingDescription` set, the
`non-staging` profile is deactivated, leading to the omission of the
`test` module which contains the `junit4-stub`. As we try to resolve
test-scoped dependencies this leads to build failure (in release).

This implements Maven plugin omissions similar to how it's done in the
`flash` profile to evade the need for building tests and downloading
test dependencies.

Also makes the release build much faster.