refactor(github): redo GitHub workflow definitions
Seems that we have accumulated a lot of complexity, unneeded duplication
and straight up unnecessary configuration in the GitHub workflow
definitions. This simplifies, optimizes and reduces the complexity of
those.

We have a single workflow `checks.yml` that applies to both pull
requests and to release branches (`N.M.x`).

Work done is minimized by checking what is changed, and build/checks are
only run for the subset of changes. For any change in the `docs/**` we
don't run any check; backend, operator and ui-react are built only if
changes have been made within their specific directories; images are
built for any change in the backend, operator or ui-react to keep the
image set consistent; and lastly integration tests are run for any
change within the backend.