fix(ui): additional step for SOAP service and port
We used to have two screens on the "Imported Operations" step, akka
review step of the custom API client wizard. That caused issues with
error handling: the screen to specify service and port would throw an
`Error` if the provided WSDL file cannot be parsed. So in #9633 instead
of reaching this screen a notification would appear indicating of this
error situation. Unfortunately, this also changed the flow for the
OpenAPI. So when an error was generated editing the successfully parsed
OpenAPI document we would render the first screen (upload) of the wizard
and show the error notification. And in process loose any changes user
has made to the document.

This instead adds an additional step in the wizard "Additional
Configuration" to specify service/port for WSDL documents, and is
skipped for OpenAPI documents, also changing the order of the
"Imported Operations" (review) screen, so that it precedes this new
screen. This way we won't `Error`-out in trying to access nonexistent
data, as it could not be parsed.

Ref. https://issues.redhat.com/browse/ENTESB-16902